fix: move Avatar component
셋팅컴포넌트에서 컴포넌트 폴더로 위치이동
마이페이지에서도 사용예정
