+++ front/src/components/Avatar/Avatar.style.js
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+import { FlexRowStyle } from 'components/globalComponents/Container/Flex.style';
+import ButtonStyle from 'components/globalComponents/Button/Button.style';
+
+export const AvatarImage = styled.img`
+  width: 80px;
+  height: 80px;
+`;
+
+export const AvatarChoiceStyle = styled(FlexRowStyle)`
+  justify-content: center;
+`;
+
+export const LeftBtn = styled(ButtonStyle)``;
+export const RightBtn = styled(ButtonStyle)``;
